Split up machines into different files
